[Chore] #39 - 코드리뷰반영(UIButton+ 익스텐션추가)
