+++ Peekabook/Peekabook/Presentation/AddBook/VC/BarcodeViewController.swift
@@ -112,15 +112,3 @@ extension BarcodeViewController: BarcodeScannerErrorDelegate {
     print(error)
   }
 }
-
-extension UIButton {
-    func setUnderline() {
-        guard let title = title(for: .normal) else { return }
-        let attributedString = NSMutableAttributedString(string: title)
-        attributedString.addAttribute(.underlineStyle,
-            value: NSUnderlineStyle.single.rawValue,
-            range: NSRange(location: 0, length: title.count)
-        )
-        setAttributedTitle(attributedString, for: .normal)
-    }
-}
